complete stack safe implementation
