+++ packages/effect/src/ref.ts
@@ -1,43 +1,61 @@
 /*
   based on: https://github.com/rzeigler/waveguide/blob/master/src/ref.ts
-  credits to original author
  */
 
 import { FunctionN } from "fp-ts/lib/function";
 import * as T from "./";
 
 export interface Ref<A> {
+  /**
+   * Get the current value of the Ref
+   */
   readonly get: T.Effect<T.NoEnv, never, A>;
-
+  /**
+   * Set the current value of the ref
+   * @param a
+   */
   set(a: A): T.Effect<T.NoEnv, never, A>;
-
+  /**
+   * Update the current value of the ref with a function.
+   * Produces the new value
+   * @param f
+   */
   update(f: FunctionN<[A], A>): T.Effect<T.NoEnv, never, A>;
-
+  /**
+   * Update the current value of a ref with a function.
+   *
+   * This function may return a second value of type B that will be produced on complete
+   * @param f
+   */
   modify<B>(f: FunctionN<[A], readonly [B, A]>): T.Effect<T.NoEnv, never, B>;
 }
 
-export function makeRef<A>(initial: A): T.Effect<T.NoEnv, never, Ref<A>> {
-  return T.fromIO(() => {
+/**
+ * Creates an IO that will allocate a Ref.
+ * Curried form of makeRef_ to allow for inference on the initial type
+ */
+export const makeRef = <A>(initial: A): T.Effect<T.NoEnv, never, Ref<A>> =>
+  T.sync(() => {
     let value = initial;
 
-    const get = T.fromIO(() => value);
+    const get = T.sync(() => value);
 
     const set = (a: A): T.Effect<T.NoEnv, never, A> =>
-      T.fromIO(() => {
+      T.sync(() => {
         const prev = value;
         value = a;
         return prev;
       });
 
     const update = (f: FunctionN<[A], A>): T.Effect<T.NoEnv, never, A> =>
-      T.fromIO(() => {
+      T.sync(() => {
         return (value = f(value));
       });
 
     const modify = <B>(
       f: FunctionN<[A], readonly [B, A]>
     ): T.Effect<T.NoEnv, never, B> =>
-      T.fromIO(() => {
+      T.sync(() => {
         const [b, a] = f(value);
         value = a;
         return b;
@@ -50,4 +68,3 @@ export function makeRef<A>(initial: A): T.Effect<T.NoEnv, never, Ref<A>> {
       modify
     };
   });
-}
