+++ packages/effect/tests/Ref.test.ts
@@ -1,5 +1,5 @@
-import * as T from "../../src/effect";
-import * as R from "../../src/effect/ref";
+import * as T from "../src";
+import * as R from "../src/ref";
 import * as assert from "assert";
 import { Do } from "fp-ts-contrib/lib/Do";
